Turn 'metric_f' into 'metric'
After talking to Kyle in IRC, it's pretty clear that the metric field needing to
be a float is an artifact of protocol buffers being strongly typed.  In the
future he'll be adding the ability to send an int instead of a float.

For dynamically typed languages like Python and Ruby, the Riemann client
libraries should just take a "metric" value.  They should handle the figuring
out of which type to send themselves, so the user doesn't have to think about
it.

This changeset renames 'metric_f' to 'metric' to be forwards-compatible, and
pulls out the setting of the metric value into a separate if block.  It also
will make it easier to add support for integer metrics once those land in
Riemann.